Resolves Issue #400.
Past: The active class added some extra margin to the active item, which was overriding the property margin for a:first-child { margin-left: 0}, a:last-child {margin-right: 0}.
Now: Resolved by doing this with relative positioning instead and increased the shift from 0.1% to 0.2%.

Also added classes to make this component usable with 2 to 10 steps, by adding the class nav-process-steps-N (where N is the number of steps). Which is documentent in the code example.